+++ test/index.test.js
@@ -1,6 +1,43 @@
 'use strict';
 
-import Validator from '../src/index';
+import Validator                from '../src/index';
+import vAlphaNumSpaces          from '../src/functions/vAlphaNumSpaces';
+import vAlphaNumSpacesMultiline from '../src/functions/vAlphaNumSpacesMultiline';
+import vArray                   from '../src/functions/vArray';
+import vArrayNe                 from '../src/functions/vArrayNe';
+import vBetween                 from '../src/functions/vBetween';
+import vBetweenInclusive        from '../src/functions/vBetweenInclusive';
+import vBoolean                 from '../src/functions/vBoolean';
+import vColorHex                from '../src/functions/vColorHex';
+import vDate                    from '../src/functions/vDate';
+import vDateString              from '../src/functions/vDateString';
+import vEmail                   from '../src/functions/vEmail';
+import vEqualTo                 from '../src/functions/vEqualTo';
+import vGeoLatitude             from '../src/functions/vGeoLatitude';
+import vGeoLongitude            from '../src/functions/vGeoLongitude';
+import vGreaterThan             from '../src/functions/vGreaterThan';
+import vGreaterThanOrEqual      from '../src/functions/vGreaterThanOrEqual';
+import vGuid                    from '../src/functions/vGuid';
+import vIn                      from '../src/functions/vIn';
+import vInteger                 from '../src/functions/vInteger';
+import vLessThan                from '../src/functions/vLessThan';
+import vLessThanOrEqual         from '../src/functions/vLessThanOrEqual';
+import vMax                     from '../src/functions/vMax';
+import vMin                     from '../src/functions/vMin';
+import vNumber                  from '../src/functions/vNumber';
+import vObject                  from '../src/functions/vObject';
+import vObjectNe                from '../src/functions/vObjectNe';
+import vPhone                   from '../src/functions/vPhone';
+import vRequired                from '../src/functions/vRequired';
+import vSize                    from '../src/functions/vSize';
+import vString                  from '../src/functions/vString';
+import vStringNe                from '../src/functions/vStringNe';
+import vSysMac                  from '../src/functions/vSysMac';
+import vSysIPv4                 from '../src/functions/vSysIPv4';
+import vSysIPv6                 from '../src/functions/vSysIPv6';
+import vSysIPv4_or_v6           from '../src/functions/vSysIPv4_or_v6';
+import vUrl                     from '../src/functions/vUrl';
+import vUrlNoQuery              from '../src/functions/vUrlNoQuery';
 
 const chai = require('chai');
 const spies = require('chai-spies');
@@ -33,7 +70,7 @@ describe("Validator - Core", () => {
 
 //  Validate that it has all the necessary functions
 
-    it ('should instantiate to a validator object', () => {
+    it ('[general] should instantiate to a validator object', () => {
         const validator = new Validator({});
 
         //  It should have a validate function on its instance
@@ -69,75 +106,7 @@ describe("Validator - Core", () => {
         should.not.exist(Validator.errors);
     });
 
-//  Validate that the returned value from validate(...) has the correct format
-
-    it ('should return a properly formatted evaluation object', () => {
-        const evaluation = (new Validator({a : 'number'})).validate(subject);
-
-        //  Evaluate object
-        assert.typeOf(evaluation, 'Object');
-        expect(Object.keys(evaluation)).to.deep.equal(['is_valid', 'errors']);
-
-        //  Evaluate object structure : is_valid
-        should.exist(evaluation.is_valid);
-        assert.typeOf(evaluation.is_valid, 'Boolean');
-
-        //  Evaluate object structure : errors
-        should.exist(evaluation.errors);
-        assert.typeOf(evaluation.errors, 'Object');
-        expect(evaluation.errors).to.deep.equal({a: []});
-    });
-
-    it ('should have errors where the object contains a msg and a params property', () => {
-        const evaluation = (new Validator({c : 'number'})).validate(subject);
-
-        expect(evaluation.errors).to.deep.equal({
-            c: [{msg: 'number', params: []}],
-        });
-    });
-
-    it ('should have errors where the key is the rule that failed', () => {
-        const evaluation = (new Validator({c : 'number'})).validate(subject);
-        expect(evaluation.errors.c[0].msg).to.eql('number');
-    });
-
-    it ('should have errors where the params object is empty if no param was passed', () => {
-        const evaluation = (new Validator({c : 'number'})).validate(subject);
-        expect(evaluation.errors.c[0].params).to.deep.equal([]);
-    });
-
-    it ('should have errors where the params object contains the passed param if passed', () => {
-        const evaluation = (new Validator({a : 'greater_than:150'})).validate(subject);
-        expect(evaluation.errors.a[0].params).to.deep.equal(['150']);
-    });
-
-//  Validate functionality
-
-    it ('should validate to true if no data was passed and no rules were set up', () => {
-        const validator = new Validator({});
-        const evaluation = validator.validate();
-        expect(evaluation.is_valid).to.eql(true);
-    });
-
-    it ('should remember the validity of the last validation run', () => {
-        const validator = new Validator({a : 'number' });
-
-        validator.validate({a: 'foo' });
-
-        expect(validator.is_valid).to.eql(false);
-        expect(validator.errors).to.deep.equal({
-            a: [{msg: 'number', params: []}],
-        });
-
-        validator.validate({a: 10 });
-
-        expect(validator.is_valid).to.eql(true);
-        expect(validator.errors).to.deep.equal({
-            a: [],
-        });
-    });
-
-    it ('should throw a type error when passed wrong configuration options', () => {
+    it ('[general] should throw a type error when passed wrong configuration options', () => {
         expect(function () {
             const v = new Validator();
         }).to.throw();
@@ -171,15 +140,83 @@ describe("Validator - Core", () => {
         }).to.throw();
     });
 
+//  Validate that the returned value from validate(...) has the correct format
+
+    it ('[validate] should return a properly formatted evaluation object', () => {
+        const evaluation = (new Validator({a : 'number'})).validate(subject);
+
+        //  Evaluate object
+        assert.typeOf(evaluation, 'Object');
+        expect(Object.keys(evaluation)).to.deep.equal(['is_valid', 'errors']);
+
+        //  Evaluate object structure : is_valid
+        should.exist(evaluation.is_valid);
+        assert.typeOf(evaluation.is_valid, 'Boolean');
+
+        //  Evaluate object structure : errors
+        should.exist(evaluation.errors);
+        assert.typeOf(evaluation.errors, 'Object');
+        expect(evaluation.errors).to.deep.equal({a: []});
+    });
+
+    it ('[validate] should have errors where the object contains a msg and a params property', () => {
+        const evaluation = (new Validator({c : 'number'})).validate(subject);
+
+        expect(evaluation.errors).to.deep.equal({
+            c: [{msg: 'number', params: []}],
+        });
+    });
+
+    it ('[validate] should have errors where the key is the rule that failed', () => {
+        const evaluation = (new Validator({c : 'number'})).validate(subject);
+        expect(evaluation.errors.c[0].msg).to.eql('number');
+    });
+
+    it ('[validate] should have errors where the params object is empty if no param was passed', () => {
+        const evaluation = (new Validator({c : 'number'})).validate(subject);
+        expect(evaluation.errors.c[0].params).to.deep.equal([]);
+    });
+
+    it ('[validate] should have errors where the params object contains the passed param if passed', () => {
+        const evaluation = (new Validator({a : 'greater_than:150'})).validate(subject);
+        expect(evaluation.errors.a[0].params).to.deep.equal(['150']);
+    });
+
+//  Validate functionality
+
+    it ('[validate] should validate to true if no data was passed and no rules were set up', () => {
+        const validator = new Validator({});
+        const evaluation = validator.validate();
+        expect(evaluation.is_valid).to.eql(true);
+    });
+
+    it ('[validate] should remember the validity of the last validation run', () => {
+        const validator = new Validator({a : 'number' });
+
+        validator.validate({a: 'foo' });
+
+        expect(validator.is_valid).to.eql(false);
+        expect(validator.errors).to.deep.equal({
+            a: [{msg: 'number', params: []}],
+        });
+
+        validator.validate({a: 10 });
+
+        expect(validator.is_valid).to.eql(true);
+        expect(validator.errors).to.deep.equal({
+            a: [],
+        });
+    });
+
 //  Flag: Sometimes
 
-    it ('sometimes flag(?) should validate correctly if set and no value is passed', () => {
+    it ('[flags] sometimes flag(?) should validate correctly if set and no value is passed', () => {
         const evaluation = (new Validator({a: '?number'})).validate({});
         expect(evaluation.is_valid).to.eql(true);
         expect(evaluation.errors.a).to.deep.equal([]);
     });
 
-    it ('sometimes flag (?) should not interfere with other validation rules', () => {
+    it ('[flags] sometimes flag (?) should not interfere with other validation rules', () => {
         const evaluation = (new Validator({
             a: '?number',
             b: 'number|less_than:20',
@@ -197,32 +234,32 @@ describe("Validator - Core", () => {
 
 //  Flag: Parameter
 
-    it ('parameter flag (<...>) should allow link to passed parameter', () => {
+    it ('[flags] parameter flag (<...>) should allow link to passed parameter', () => {
         const evaluation = (new Validator({a: 'equal_to:<foo>'})).validate({a: 'hello', foo: 'hello'});
         expect(evaluation.is_valid).to.eql(true);
         expect(evaluation.errors.a).to.deep.equal([]);
     });
 
-    it ('parameter flag (<...>) should fail if parameter is not passed', () => {
+    it ('[flags] parameter flag (<...>) should fail if parameter is not passed', () => {
         const evaluation = (new Validator({a: 'equal_to:<foo>'})).validate({a: 'hello', foobar: 'hello'});
         expect(evaluation.is_valid).to.eql(false);
         expect(evaluation.errors.a).to.deep.equal([{msg:'equal_to', params: [undefined]}]);
     });
 
-    it ('parameter flag (<...>) should allow multiple parameters inside the same ruleset', () => {
+    it ('[flags] parameter flag (<...>) should allow multiple parameters inside the same ruleset', () => {
         const evaluation = (new Validator({a: 'between:<min>,<max>'})).validate({a: 5, min: 3, max: 7});
         expect(evaluation.is_valid).to.eql(true);
         expect(evaluation.errors.a).to.deep.equal([]);
     });
 
-    it ('parameter flag (<...>) should allow multiple parameters inside the same config', () => {
+    it ('[flags] parameter flag (<...>) should allow multiple parameters inside the same config', () => {
         const evaluation = (new Validator({a: 'in:<arr1>', b: 'in:<arr2>'})).validate({a: 1, b: 2, arr1: [1, 3, 5], arr2: [2, 4, 6]});
         expect(evaluation.is_valid).to.eql(true);
         expect(evaluation.errors.a).to.deep.equal([]);
         expect(evaluation.errors.b).to.deep.equal([]);
     });
 
-    it ('parameter flag (<...>) should allow the same parameter on multiple rules inside the same config', () => {
+    it ('[flags] parameter flag (<...>) should allow the same parameter on multiple rules inside the same config', () => {
         const evaluation = (new Validator({a: 'in:<arr1>', b: 'in:<arr1>'})).validate({a: 1, b: 2, arr1: [1, 2, 3]});
         expect(evaluation.is_valid).to.eql(true);
         expect(evaluation.errors.a).to.deep.equal([]);
@@ -231,13 +268,13 @@ describe("Validator - Core", () => {
 
 //  Flag: Not
 
-    it ('not flag (!) should validate correct if set and no value is passed', () => {
+    it ('[flags] not flag (!) should validate correct if set and no value is passed', () => {
         const evaluation = (new Validator({a: '!number'})).validate({});
         expect(evaluation.is_valid).to.eql(false);
         expect(evaluation.errors.a).to.deep.equal([{msg: 'not_number', params: []}]);
     });
 
-    it ('not flag (!) should validate correct if set and value is passed when using standard rule', () => {
+    it ('[flags] not flag (!) should validate correct if set and value is passed when using standard rule', () => {
         const evaluation = (new Validator({a: '!number'})).validate({a: 'hello'});
         expect(evaluation.is_valid).to.eql(true);
         expect(evaluation.errors.a).to.deep.equal([]);
@@ -255,13 +292,13 @@ describe("Validator - Core", () => {
         expect(evaluation4.errors.a).to.deep.equal([{msg: 'not_between', params: ["5", "10"]}]);
     });
 
-    it ('not flag (!) should validate correct if set and value is passed when using parameter flag', () => {
+    it ('[flags] not flag (!) should validate correct if set and value is passed when using parameter flag', () => {
         const evaluation = (new Validator({a: '!equal_to:<foo>'})).validate({a: 'hello', foo: 'hello'});
         expect(evaluation.is_valid).to.eql(false);
         expect(evaluation.errors.a).to.deep.equal([{msg: 'not_equal_to', params: ['hello']}]);
     });
 
-    it ('not flag (!) should validate correct if set and value is passed when using multiple validation rules', () => {
+    it ('[flags] not flag (!) should validate correct if set and value is passed when using multiple validation rules', () => {
         const evaluation = (new Validator({a: '!string|!equal_to:<foo>'})).validate({a: 'foo', foo: 'hello'});
         expect(evaluation.is_valid).to.eql(false);
         expect(evaluation.errors.a).to.deep.equal([
@@ -278,7 +315,7 @@ describe("Validator - Core", () => {
 
 //  Iterable
 
-    it ('iterable config: Should throw if passed an invalid iterable config during rule creation', () => {
+    it ('[iterable] config: Should throw if passed an invalid iterable config during rule creation', () => {
         expect(function () {
             new Validator({a: '[string'});
         }).to.throw;
@@ -292,53 +329,53 @@ describe("Validator - Core", () => {
         }).to.throw;
     });
 
-    it ('iterable config: Should not throw if passed an iterable config during rule creation', () => {
+    it ('[iterable] config: Should not throw if passed an iterable config during rule creation', () => {
         expect(function () {
             new Validator({a: '[]string'});
         }).to.not.throw;
     });
 
-    it ('iterable config: Should return invalid when passing a non-array to an iterable', () => {
+    it ('[iterable] config: Should return invalid when passing a non-array to an iterable', () => {
         const validator = new Validator({a: '[]string'});
         let evaluation = validator.validate({a: 'hello'});
         expect(evaluation.is_valid).to.eql(false);
         expect(evaluation.errors).to.deep.equal({a: [{msg: 'iterable', params: []}]});
     });
 
-    it ('iterable config: Should return valid when passing an empty array to an iterable', () => {
+    it ('[iterable] config: Should return valid when passing an empty array to an iterable', () => {
         const validator = new Validator({a: '[]string'});
         let evaluation = validator.validate({a: []});
         expect(evaluation.is_valid).to.eql(true);
     });
 
-    it ('iterable unique config: Should return valid when the passed array is unique', () => {
+    it ('[iterable] unique config: Should return valid when the passed array is unique', () => {
         const validator = new Validator({a: '[unique]string'});
         let evaluation = validator.validate({a: ['a', 'b', 'c']});
         expect(evaluation.is_valid).to.eql(true);
     });
 
-    it ('iterable unique config: Should return invalid when the passed array is not unique', () => {
+    it ('[iterable] unique config: Should return invalid when the passed array is not unique', () => {
         const validator = new Validator({a: '[unique]string'});
         let evaluation = validator.validate({a: ['a', 'b', 'a']});
         expect(evaluation.is_valid).to.eql(false);
         expect(evaluation.errors).to.deep.equal({a: [{msg: 'iterable_unique', params: []}]});
     });
 
-    it ('iterable unique config: Should return invalid when the passed array is not unique when using numbers', () => {
+    it ('[iterable] unique config: Should return invalid when the passed array is not unique when using numbers', () => {
         const validator = new Validator({a: '[unique]number'});
         let evaluation = validator.validate({a: [1, 2, 2]});
         expect(evaluation.is_valid).to.eql(false);
         expect(evaluation.errors).to.deep.equal({a: [{msg: 'iterable_unique', params: []}]});
     });
 
-    it ('iterable unique config: Should return invalid when the passed array is not unique when using objects', () => {
+    it ('[iterable] unique config: Should return invalid when the passed array is not unique when using objects', () => {
         const validator = new Validator({a: '[unique]object'});
         let evaluation = validator.validate({a: [{a: 1}, {b: 2}, {b: 2}]});
         expect(evaluation.is_valid).to.eql(false);
         expect(evaluation.errors).to.deep.equal({a: [{msg: 'iterable_unique', params: []}]});
     });
 
-    it ('iterable unique config: Should return invalid when the passed array is both not unique and doesnt match rules', () => {
+    it ('[iterable] unique config: Should return invalid when the passed array is both not unique and doesnt match rules', () => {
         const validator = new Validator({a: '[unique]integer'});
         let evaluation = validator.validate({a: [1, 2, 'a', 2]});
         expect(evaluation.is_valid).to.eql(false);
@@ -348,7 +385,7 @@ describe("Validator - Core", () => {
         ]});
     });
 
-    it ('iterable unique config: Should return invalid when the passed array is both not unique and doesnt match rules, and should only insert uniqueness invalidity once', () => {
+    it ('[iterable] unique config: Should return invalid when the passed array is both not unique and doesnt match rules, and should only insert uniqueness invalidity once', () => {
         const validator = new Validator({a: '[unique]integer'});
         let evaluation = validator.validate({a: [1, 2, 'a', 2, 2, 2.2]});
         expect(evaluation.is_valid).to.eql(false);
@@ -359,33 +396,33 @@ describe("Validator - Core", () => {
         ]});
     });
 
-    it ('iterable max config: Should return valid when within boundaries', () => {
+    it ('[iterable] max config: Should return valid when within boundaries', () => {
         const validator = new Validator({a: '[max:5]string'});
         let evaluation = validator.validate({a: ['hello', 'there']});
         expect(evaluation.is_valid).to.eql(true);
     });
 
-    it ('iterable max config: Should return valid when at boundary', () => {
+    it ('[iterable] max config: Should return valid when at boundary', () => {
         const validator = new Validator({a: '[max:5]string'});
         let evaluation = validator.validate({a: ['1', '2', '3', '4', '5']});
         expect(evaluation.is_valid).to.eql(true);
     });
 
-    it ('iterable max config: Should return invalid when above boundary', () => {
+    it ('[iterable] max config: Should return invalid when above boundary', () => {
         const validator = new Validator({a: '[max:5]string'});
         let evaluation = validator.validate({a: ['1', '2', '3', '4', '5', '6']});
         expect(evaluation.is_valid).to.eql(false);
         expect(evaluation.errors).to.deep.equal({a: [{msg: 'iterable_max', params: [5]}]});
     });
 
-    it ('iterable max config: Should only return iterable invalidity and not go into val evaluation when above boundary', () => {
+    it ('[iterable] max config: Should only return iterable invalidity and not go into val evaluation when above boundary', () => {
         const validator = new Validator({a: '[max:5]string'});
         let evaluation = validator.validate({a: ['1', '2', '3', '4', 5, '6']});
         expect(evaluation.is_valid).to.eql(false);
         expect(evaluation.errors).to.deep.equal({a: [{msg: 'iterable_max', params: [5]}]});
     });
 
-    it ('iterable max config: Should return iterable invalidity and go into val evaluation when at or below boundary', () => {
+    it ('[iterable] max config: Should return iterable invalidity and go into val evaluation when at or below boundary', () => {
         const validator = new Validator({a: '[max:5]string'});
         let evaluation = validator.validate({a: ['1', false, false, '2', false]});
         expect(evaluation.is_valid).to.eql(false);
@@ -402,33 +439,33 @@ describe("Validator - Core", () => {
         ]});
     });
 
-    it ('iterable min config: Should return valid when within boundaries', () => {
+    it ('[iterable] min config: Should return valid when within boundaries', () => {
         const validator = new Validator({a: '[min:5]string'});
         let evaluation = validator.validate({a: ['hello', 'there', 'this', 'is', 'cool', 'right']});
         expect(evaluation.is_valid).to.eql(true);
     });
 
-    it ('iterable min config: Should return valid when at boundary', () => {
+    it ('[iterable] min config: Should return valid when at boundary', () => {
         const validator = new Validator({a: '[min:5]string'});
         let evaluation = validator.validate({a: ['hello', 'there', 'this', 'is', 'cool']});
         expect(evaluation.is_valid).to.eql(true);
     });
 
-    it ('iterable min config: Should return invalid when below boundary', () => {
+    it ('[iterable] min config: Should return invalid when below boundary', () => {
         const validator = new Validator({a: '[min:3]string'});
         let evaluation = validator.validate({a: ['1', '2']});
         expect(evaluation.is_valid).to.eql(false);
         expect(evaluation.errors).to.deep.equal({a: [{msg: 'iterable_min', params: [3]}]});
     });
 
-    it ('iterable min config: Should only return iterable invalidity and not go into val evaluation when below boundary', () => {
+    it ('[iterable] min config: Should only return iterable invalidity and not go into val evaluation when below boundary', () => {
         const validator = new Validator({a: '[min:4]string'});
         let evaluation = validator.validate({a: ['1', false]});
         expect(evaluation.is_valid).to.eql(false);
         expect(evaluation.errors).to.deep.equal({a: [{msg: 'iterable_min', params: [4]}]});
     });
 
-    it ('iterable min config: Should return iterable invalidity and go into val evaluation when at or above boundary', () => {
+    it ('[iterable] min config: Should return iterable invalidity and go into val evaluation when at or above boundary', () => {
         const validator = new Validator({a: '[min:4]string'});
         let evaluation = validator.validate({a: ['1', false, false, '2']});
         expect(evaluation.is_valid).to.eql(false);
@@ -445,53 +482,53 @@ describe("Validator - Core", () => {
         ]});
     });
 
-    it ('iterable max+min config: Should return valid when within boundaries', () => {
+    it ('[iterable] max+min config: Should return valid when within boundaries', () => {
         const validator = new Validator({a: '[max:5|min:2]string'});
         let evaluation = validator.validate({a: ['hello', 'there', 'cool']});
         expect(evaluation.is_valid).to.eql(true);
     });
 
-    it ('iterable max+min config: Should return valid when at top boundary', () => {
+    it ('[iterable] max+min config: Should return valid when at top boundary', () => {
         const validator = new Validator({a: '[max:5|min:2]string'});
         let evaluation = validator.validate({a: ['1', '2', '3', '4', '5']});
         expect(evaluation.is_valid).to.eql(true);
     });
 
-    it ('iterable max+min config: Should return invalid when above top boundary', () => {
+    it ('[iterable] max+min config: Should return invalid when above top boundary', () => {
         const validator = new Validator({a: '[max:5|min:2]string'});
         let evaluation = validator.validate({a: ['1', '2', '3', '4', '5', '6']});
         expect(evaluation.is_valid).to.eql(false);
         expect(evaluation.errors).to.deep.equal({a: [{msg: 'iterable_max', params: [5]}]});
     });
 
-    it ('iterable max+min config: Should return valid when at bottom boundary', () => {
+    it ('[iterable] max+min config: Should return valid when at bottom boundary', () => {
         const validator = new Validator({a: '[max:5|min:2]string'});
         let evaluation = validator.validate({a: ['1', '2']});
         expect(evaluation.is_valid).to.eql(true);
     });
 
-    it ('iterable max+min config: Should return invalid when below bottom boundary', () => {
+    it ('[iterable] max+min config: Should return invalid when below bottom boundary', () => {
         const validator = new Validator({a: '[max:5|min:2]string'});
         let evaluation = validator.validate({a: ['1']});
         expect(evaluation.is_valid).to.eql(false);
         expect(evaluation.errors).to.deep.equal({a: [{msg: 'iterable_min', params: [2]}]});
     });
 
-    it ('iterable max+min config: Should only return iterable invalidity and not go into val evaluation when above boundary', () => {
+    it ('[iterable] max+min config: Should only return iterable invalidity and not go into val evaluation when above boundary', () => {
         const validator = new Validator({a: '[max:5|min:2]string'});
         let evaluation = validator.validate({a: ['1', '2', '3', '4', 5, '6']});
         expect(evaluation.is_valid).to.eql(false);
         expect(evaluation.errors).to.deep.equal({a: [{msg: 'iterable_max', params: [5]}]});
     });
 
-    it ('iterable max+min config: Should only return iterable invalidity and not go into val evaluation when below boundary', () => {
+    it ('[iterable] max+min config: Should only return iterable invalidity and not go into val evaluation when below boundary', () => {
         const validator = new Validator({a: '[max:5|min:2]string'});
         let evaluation = validator.validate({a: [5]});
         expect(evaluation.is_valid).to.eql(false);
         expect(evaluation.errors).to.deep.equal({a: [{msg: 'iterable_min', params: [2]}]});
     });
 
-    it ('iterable max+min config: Should return iterable invalidity and go into val evaluation when at or below boundary', () => {
+    it ('[iterable] max+min config: Should return iterable invalidity and go into val evaluation when at or below boundary', () => {
         const validator = new Validator({a: '[max:5|min:2]string'});
         let evaluation = validator.validate({a: ['1', false, false, '2', false]});
         expect(evaluation.is_valid).to.eql(false);
@@ -514,7 +551,7 @@ describe("Validator - Core", () => {
         ]});
     });
 
-    it ('iterable: Should allow validating an array of values', () => {
+    it ('[iterable] Should allow validating an array of values', () => {
         const validator = new Validator({a: '[]string'});
         let evaluation = validator.validate({a: ['hello', 'there']});
         expect(evaluation.is_valid).to.eql(true);
@@ -528,7 +565,7 @@ describe("Validator - Core", () => {
         expect(evaluation.errors.a).to.deep.equal([{msg: 'string', idx: 1, params: []}, {msg: 'string', idx: 2, params: []}]);
     });
 
-    it ('iterable: Should allow validating an array of values with a \'?\' sometimes flag', () => {
+    it ('[iterable] Should allow validating an array of values with a \'?\' sometimes flag', () => {
         const validator = new Validator({a: '?[]string'});
         let evaluation = validator.validate({a: ['hello', 'there']});
         expect(evaluation.is_valid).to.eql(true);
@@ -541,7 +578,7 @@ describe("Validator - Core", () => {
         expect(evaluation.errors.a).to.deep.equal([{msg: 'string', idx: 1, params: []}, {msg: 'string', idx: 3, params: []}]);
     });
 
-    it ('iterable: Should allow validating an array of values with a \'?\' sometimes flag and parameter pass', () => {
+    it ('[iterable] Should allow validating an array of values with a \'?\' sometimes flag and parameter pass', () => {
         const validator = new Validator({a: '?[]string|in:<genders>'});
         let evaluation = validator.validate({a: ['male', 'male', 'female', 'male', 'female'], genders: ['male', 'female', 'other']});
         expect(evaluation.is_valid).to.eql(true);
@@ -553,7 +590,7 @@ describe("Validator - Core", () => {
         expect(evaluation.is_valid).to.eql(false);
     });
 
-    it ('iterable: Should allow validating an array of values with a \'?\' sometimes flag, uniqueness and parameter pass', () => {
+    it ('[iterable] Should allow validating an array of values with a \'?\' sometimes flag, uniqueness and parameter pass', () => {
         const validator = new Validator({a: '?[unique]string|in:<genders>'});
         let evaluation = validator.validate({a: ['male', 'male', 'female', 'male', 'female'], genders: ['male', 'female', 'other']});
         expect(evaluation.is_valid).to.eql(false);
@@ -561,7 +598,7 @@ describe("Validator - Core", () => {
         expect(evaluation.is_valid).to.eql(true);
     });
 
-    it ('iterable: Should allow validating an array of values with a \'!\' not flag and parameter pass', () => {
+    it ('[iterable] Should allow validating an array of values with a \'!\' not flag and parameter pass', () => {
         const validator = new Validator({a: '[]string|!in:<genders>'});
         let evaluation = validator.validate({a: ['bob', 'bob', 'john', 'bob', 'john'], genders: ['male', 'female', 'other']});
         expect(evaluation.is_valid).to.eql(true);
@@ -581,7 +618,7 @@ describe("Validator - Core", () => {
         ]});
     });
 
-    it ('iterable: Should allow validating an array of values with a \'!\' not flag, uniqueness and parameter pass', () => {
+    it ('[iterable] Should allow validating an array of values with a \'!\' not flag, uniqueness and parameter pass', () => {
         const validator = new Validator({a: '[unique]string|!in:<genders>'});
         let evaluation = validator.validate({a: ['bob', 'john'], genders: ['male', 'female', 'other']});
         expect(evaluation.is_valid).to.eql(true);
@@ -603,15 +640,86 @@ describe("Validator - Core", () => {
         ]});
     });
 
-//  Extend functionality
+//  Static: rules getter
 
-    it ('extend functionality should throw if not provided anything', () => {
+    it ('[rules] should return the configured rules on the Validator as an object', () => {
+        assert.typeOf(Validator.rules, 'Object');
+    });
+
+    it ('[rules] should return a correct kv-map of configured rules', () => {
+        expect(Validator.rules).to.deep.equal({
+            alpha_num_spaces            : vAlphaNumSpaces,
+            alpha_num_spaces_multiline  : vAlphaNumSpacesMultiline,
+            array                       : vArray,
+            array_ne                    : vArrayNe,
+            between                     : vBetween,
+            between_inc                 : vBetweenInclusive,
+            boolean                     : vBoolean,
+            color_hex                   : vColorHex,
+            date                        : vDate,
+            date_string                 : vDateString,
+            email                       : vEmail,
+            equal_to                    : vEqualTo,
+            geo_latitude                : vGeoLatitude,
+            geo_longitude               : vGeoLongitude,
+            greater_than                : vGreaterThan,
+            greater_than_or_equal       : vGreaterThanOrEqual,
+            guid                        : vGuid,
+            in                          : vIn,
+            integer                     : vInteger,
+            less_than                   : vLessThan,
+            less_than_or_equal          : vLessThanOrEqual,
+            max                         : vMax,
+            min                         : vMin,
+            number                      : vNumber,
+            object                      : vObject,
+            object_ne                   : vObjectNe,
+            phone                       : vPhone,
+            required                    : vRequired,
+            size                        : vSize,
+            string                      : vString,
+            string_ne                   : vStringNe,
+            sys_mac                     : vSysMac,
+            sys_ipv4                    : vSysIPv4,
+            sys_ipv6                    : vSysIPv6,
+            sys_ipv4_or_v6              : vSysIPv4_or_v6,
+            url                         : vUrl,
+            url_noquery                 : vUrlNoQuery,
+            gt                          : vGreaterThan,
+            gte                         : vGreaterThanOrEqual,
+            lt                          : vLessThan,
+            lte                         : vLessThanOrEqual,
+            eq                          : vEqualTo,
+        });
+    });
+
+    it ('[rules] should return a immutable frozen version of the configured rules', () => {
+        const rules = Validator.rules;
+        expect(Object.isFrozen(rules)).to.eql(true);
+        let e = false;
+        try {
+            rules.hack = true;
+        } catch (err) {
+            e = err;
+        }
+        assert.typeOf(e, 'Error');
+    });
+
+    it ('[rules] should take into account extended rules', () => {
+        const fn = (val) => true;
+        Validator.extend('myfunction', fn);
+        expect(Validator.rules.myfunction).to.eql(fn);
+    });
+
+//  Static: Extend functionality
+
+    it ('[extend] functionality should throw if not provided anything', () => {
         expect(function () {
             Validator.extend();
         }).to.throw;
     });
 
-    it ('extend functionality should throw if not provided a string name', () => {
+    it ('[extend] functionality should throw if not provided a string name', () => {
         for (const el of str_tests) {
             expect(function () {
                 Validator.extend(el, val => true);
@@ -619,7 +727,7 @@ describe("Validator - Core", () => {
         }
     });
 
-    it ('extend functionality should throw if provided a empty (or empty after trimming) string name', () => {
+    it ('[extend] functionality should throw if provided a empty (or empty after trimming) string name', () => {
         for (const el of ['', ' ', '    ']) {
             expect(function () {
                 Validator.extend(el, val => true);
@@ -627,7 +735,7 @@ describe("Validator - Core", () => {
         }
     });
 
-    it ('extend functionality should throw if not provided a function', () => {
+    it ('[extend] functionality should throw if not provided a function', () => {
         for (const el of fn_tests) {
             expect(function () {
                 Validator.extend('rule', el);
@@ -635,7 +743,7 @@ describe("Validator - Core", () => {
         }
     });
 
-    it ('extend functionality should work', () => {
+    it ('[extend] functionality should work', () => {
         Validator.extend('trick', function (val, p1) {
             return p1 === 'treat' ? (val === 'trick') : (p1 === 'trick' ? val === 'treat' : false);
         });
@@ -645,7 +753,7 @@ describe("Validator - Core", () => {
         expect(evaluation.errors.a).to.deep.equal([]);
     });
 
-    it ('extend functionality should work with multiple parameters', () => {
+    it ('[extend] functionality should work with multiple parameters', () => {
         Validator.extend('sum', function (val, p1, p2) {
             return val === (p1 + p2);
         });
@@ -655,7 +763,7 @@ describe("Validator - Core", () => {
         expect(evaluation.errors.a).to.deep.equal([]);
     });
 
-    it ('extend functionality should work across multiple instances', () => {
+    it ('[extend] functionality should work across multiple instances', () => {
         Validator.extend('double', function (val, p1, p2) {
             return val === (p1 * 2);
         });
@@ -671,7 +779,7 @@ describe("Validator - Core", () => {
         expect(evaluation2.errors.a).to.deep.equal([]);
     });
 
-    it ('extend functionality should allow redefining the same validity function', () => {
+    it ('[extend] functionality should allow redefining the same validity function', () => {
         Validator.extend('ismyname', function (val, p1, p2) {return val === 'peter';});
 
         //  Evaluation
@@ -690,13 +798,13 @@ describe("Validator - Core", () => {
 
 //  Extend Multi functionality
 
-    it ('extendMulti functionality should throw if not provided anything', () => {
+    it ('[extendMulti] functionality should throw if not provided anything', () => {
         expect(function () {
             Validator.extend();
         }).to.throw;
     });
 
-    it ('extendMulti functionality should throw if not provided an object', () => {
+    it ('[extendMulti] functionality should throw if not provided an object', () => {
         for (const el of obj_tests) {
             expect(function () {
                 Validator.extendMulti(el);
@@ -704,13 +812,13 @@ describe("Validator - Core", () => {
         }
     });
 
-    it ('extendMulti functionality should not throw if provided an empty object', () => {
+    it ('[extendMulti] functionality should not throw if provided an empty object', () => {
         expect(function () {
             Validator.extendMulti({});
         }).to.not.throw;
     });
 
-    it ('extendMulti functionality should throw if provided an object where certain values do not have a function', () => {
+    it ('[extendMulti] functionality should throw if provided an object where certain values do not have a function', () => {
         for (const el of fn_tests) {
             expect(function () {
                 Validator.extendMulti({
@@ -720,7 +828,7 @@ describe("Validator - Core", () => {
         }
     });
 
-    it ('extendMulti functionality should work', () => {
+    it ('[extendMulti] functionality should work', () => {
         Validator.extendMulti({
             trick: function (val, p1) {
                 return p1 === 'treat' ? (val === 'trick') : (p1 === 'trick' ? val === 'treat' : false);
@@ -732,7 +840,7 @@ describe("Validator - Core", () => {
         expect(evaluation.errors.a).to.deep.equal([]);
     });
 
-    it ('extendMulti functionality should work with multiple parameters', () => {
+    it ('[extendMulti] functionality should work with multiple parameters', () => {
         Validator.extendMulti({
             sum: function (val, p1, p2) {
                 return val === (p1 + p2);
@@ -744,7 +852,7 @@ describe("Validator - Core", () => {
         expect(evaluation.errors.a).to.deep.equal([]);
     });
 
-    it ('extendMulti functionality should work across multiple instances', () => {
+    it ('[extendMulti] functionality should work across multiple instances', () => {
         Validator.extendMulti({
             double: function (val, p1, p2) {
                 return val === (p1 * 2);
@@ -762,7 +870,7 @@ describe("Validator - Core", () => {
         expect(evaluation2.errors.a).to.deep.equal([]);
     });
 
-    it ('extendMulti functionality should allow redefining the same validity function', () => {
+    it ('[extendMulti] functionality should allow redefining the same validity function', () => {
         Validator.extendMulti({
             ismyname: function (val, p1, p2) {return val === 'peter';},
         });
@@ -783,7 +891,7 @@ describe("Validator - Core", () => {
         expect(evaluation2.errors.name).to.deep.equal([{msg:'ismyname', params: []}]);
     });
 
-    it ('extendMulti functionality should allow defining multiple rules at the same time', () => {
+    it ('[extendMulti] functionality should allow defining multiple rules at the same time', () => {
         Validator.extendMulti({
             ismyname: function (val, p1, p2) {return val === 'peter';},
             ismyothername: function (val, p1, p2) {return val === 'josh';},
@@ -800,9 +908,9 @@ describe("Validator - Core", () => {
         expect(evaluation2.errors.name).to.deep.equal([{msg:'ismyothername', params: []}]);
     });
 
-//  Complex validations
+//  Complex Validations
 
-    it ('should be able to validate complex objects [1]', () => {
+    it ('[complexflow] should be able to validate complex objects [1]', () => {
         const evaluation = (new Validator({
             password: 'string|min:8',
             password_confirmation: 'equal_to:<password>',
@@ -816,7 +924,7 @@ describe("Validator - Core", () => {
         expect(evaluation.errors.password_confirmation).to.deep.equal([]);
     });
 
-    it ('should be able to validate complex objects [2]', () => {
+    it ('[complexflow] should be able to validate complex objects [2]', () => {
         const evaluation = (new Validator({
             first_name: 'string|alpha_num_spaces|min:2',
             last_name: 'string|alpha_num_spaces|min:2',
@@ -838,7 +946,7 @@ describe("Validator - Core", () => {
         expect(evaluation.errors.gender).to.deep.equal([]);
     });
 
-    it ('should be able to validate complex objects [3]', () => {
+    it ('[complexflow] should be able to validate complex objects [3]', () => {
         const evaluation = (new Validator({
             first_name: 'string|alpha_num_spaces|min:2',
             last_name: 'string|alpha_num_spaces|min:2',
@@ -861,7 +969,7 @@ describe("Validator - Core", () => {
         expect(evaluation.errors.gender).to.deep.equal([{msg: 'in', params: [['m', 'f', 'o']]}]);
     });
 
-    it ('should be able to validate complex multidimensional objects [4]', () => {
+    it ('[complexflow] should be able to validate complex multidimensional objects [4]', () => {
         const evaluation = (new Validator({
             address: {
                 street: 'string|alpha_num_spaces',
@@ -889,7 +997,7 @@ describe("Validator - Core", () => {
         expect(evaluation.errors.contact.email).to.deep.equal([{msg: 'email', params: []}]);
     });
 
-    it ('should be able to validate complex multidimensional objects [5]', () => {
+    it ('[complexflow] should be able to validate complex multidimensional objects [5]', () => {
         Validator.extend('is_type', val => ['type1', 'type2', 'type4'].indexOf(val) >= 0);
 
         const validator = new Validator({